requirements.txt
attrs
torch
transformers
hkkang_utils
unittest-xml-reporting
hydra-core
gitpython
wandb